refactor: 쿠폰 할인 계산 로직을 Coupon 도메인 모델 내부로 이동
